review: consolidate sensitive-data test, drop now-redundant comment
Address @christinaexyou's review on PR #1967:

- Move `test_mask_sensitive_data_honors_configured_score_threshold` from the
  standalone `tests/test_sensitive_data_detection_unit.py` into the existing
  `tests/test_sensitive_data_detection.py` (with the same `skipif` +
  `unit` + `asyncio` decorator stack used by the other tests). Delete the
  now-redundant unit file.

- Drop the four-line explanatory comment above the fixed `_get_analyzer`
  call; the code is self-explanatory once the bug is gone and the reasoning
  lives in the commit message.

Also add a `if SDD_SETUP_PRESENT: import spacy` shim so the module's
`setup_module` (which references `spacy.util.is_package` without ever
importing `spacy`) actually runs when the SDD extras are installed.
Previously every test in the file was silently skipped via the bare
`except Exception` fallthrough with "Unexpected error during setup: name
'spacy' is not defined" — making local verification impossible.

Verified locally with `pip install presidio-analyzer presidio-anonymizer
spacy && python -m spacy download en_core_web_lg`: the regression test
PASSES with the fix and FAILS (AssertionError) on a manually reverted
`actions.py`.
